Restructure the sidebar for better pointer handling (touch/mouse)
Had to be restructured completely for reasonable outcome. Looks like I nailed it! Vertically and horizontally responsive. Cheers!
