+++ src/components/ControlBox.tsx
@@ -4,9 +4,13 @@ import { SANS } from '../utils/styled'
 
 export const ControlBoxBorderRadius = 10
 
+export const DefaultControlBoxBackgroundColor = '#fcfcfc'
+
+export const ControlBoxShadow = '0 0 6px rgba(0, 0, 0, 0.08)'
+
 const ControlBox = styled.div`
-  background: #fcfcfc;
-  box-shadow: 0 0 6px rgba(0, 0, 0, 0.08);
+  background: ${DefaultControlBoxBackgroundColor};
+  box-shadow: ${ControlBoxShadow};
   border-radius: ${ControlBoxBorderRadius}px;
   font-family: ${SANS};
 `
